requirements.txt
langchain
langchain-ollama
ollama
langchain-community
langchain-huggingface
pypdf
python-docx
unstructured
tiktoken
sentence-transformers
chromadb
huggingface_hub

faiss-cpu
markdown
pandas
openpyxl
msoffcrypto-tool
beautifulsoup4 
html2text
nltk

#bs4
#chromadb
#faiss-cpu
#wikipedia

#ddgs
#duckduckgo-search
#langchain-experimental